+++ zipkin/src/main/java/com/linecorp/armeria/common/tracing/package-info.java
@@ -15,8 +15,7 @@
  */
 
 /**
- * Distributed tracing clients based on <a href="https://github.com/openzipkin/brave">Brave</a>,
- * a Java tracing library compatible with <a href="http://zipkin.io/">Zipkin</a>.
+ * Common classes for distributed tracing support based on <a href="https://github.com/openzipkin/brave">Brave</a>.
  */
 @NonNullByDefault
 package com.linecorp.armeria.common.tracing;
