--- zipkin/src/test/java/com/linecorp/armeria/internal/tracing/PingPongExtraTest.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2018 LINE Corporation
- *
- * LINE Corporation licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-
-package com.linecorp.armeria.internal.tracing;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Collections;
-
-import org.junit.Test;
-
-import brave.propagation.TraceContext;
-
-public class PingPongExtraTest {
-
-    @Test
-    public void shouldSetPongIfOnlyExtra() {
-        final PingPongExtra extra = new PingPongExtra();
-
-        final TraceContext context = TraceContext.newBuilder().traceId(1).spanId(1)
-                                                 .extra(Collections.singletonList(extra)).build();
-
-        PingPongExtra.maybeSetPong(context);
-
-        assertThat(extra.isPong()).isTrue();
-    }
-}
